[TensorExpr] Consolidate {buffer,function,tensor}.{h.cpp} in tensor.{h,cpp}.
Classes defined in these files are closely related, so it is reasonable
to have them all in one file. The change is purely a code move.

[ghstack-poisoned]
