+++ torch/csrc/jit/tensorexpr/tensor.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <torch/csrc/WindowsTorchApiMacro.h>
+#include <functional>
 #include <vector>
 
 #include <torch/csrc/jit/tensorexpr/dim_arg.h>
@@ -11,6 +12,96 @@
 namespace torch {
 namespace jit {
 namespace tensorexpr {
+
+class Function : public KernelScopedObject {
+ public:
+  Function(
+      const std::string& func_name,
+      const std::vector<const Expr*>& dims,
+      const std::vector<const Var*>& args,
+      const Expr* body)
+      // TODO: Function should not create buffers, they should be created
+      // manually before constructing a function.
+      : func_vars_({new Buf(func_name, dims, body->dtype())}),
+        dims_(dims),
+        args_(args),
+        bodies_({body}) {}
+  Function(
+      const std::vector<std::string>& func_names,
+      const std::vector<const Expr*>& dims,
+      const std::vector<const Var*>& args,
+      const std::vector<const Expr*>& bodies)
+      : func_vars_(func_names.size()),
+        dims_(dims),
+        args_(args),
+        bodies_(bodies) {
+    for (size_t i = 0; i < func_names.size(); i++) {
+      func_vars_[i] = new Buf(func_names[i], dims, bodies[i]->dtype());
+    }
+  }
+  Function(
+      const std::string& func_name,
+      Buf* func_var,
+      const std::vector<const Expr*>& dims,
+      const std::vector<const Var*>& args,
+      const Expr* body)
+      : func_vars_({func_var}), dims_(dims), args_(args), bodies_({body}) {}
+
+  size_t ndim() const {
+    return dims_.size();
+  }
+
+  const Expr* dim(size_t index) const {
+    if (index < 0 || index >= dims_.size()) {
+      throw out_of_range_index();
+    }
+
+    return dims_[index];
+  }
+  const std::vector<const Expr*>& dims() const {
+    return dims_;
+  }
+
+  const Var* arg(size_t index) const {
+    if (index < 0 || index >= args_.size()) {
+      throw out_of_range_index();
+    }
+
+    return args_[index];
+  }
+  const std::vector<const Var*>& args() const {
+    return args_;
+  }
+
+  std::vector<const Expr*> bodies() const {
+    return bodies_;
+  }
+  const Expr* body(size_t index) const {
+    if (index >= bodies_.size()) {
+      throw out_of_range_index();
+    }
+
+    return bodies_[index];
+  }
+
+  std::vector<const Buf*> func_vars() const {
+    return func_vars_;
+  }
+  const Buf* func_var(size_t index) const {
+    if (index >= func_vars_.size()) {
+      throw out_of_range_index();
+    }
+    return func_vars_[index];
+  }
+
+  Stmt* ElementStmt(size_t index);
+
+ private:
+  std::vector<const Buf*> func_vars_;
+  std::vector<const Expr*> dims_;
+  std::vector<const Var*> args_;
+  std::vector<const Expr*> bodies_;
+};
 
 class Tensor : KernelScopedObject {
  public:
@@ -70,6 +161,82 @@ class Tensor : KernelScopedObject {
   int output_index_;
   const Expr* initializer_{nullptr};
 };
+
+class Buffer {
+ public:
+  Buffer(const BufHandle& data) : data_(data.node()) {
+    if (data_->base_handle()->dtype() != kHandle) {
+      throw malformed_input("Buffer dtype must be Handle");
+    }
+
+    std::vector<ExprHandle> stride_handles(ndim());
+    for (int i = (int)ndim() - 1; i >= 0; i--) {
+      if (i == ndim() - 1) {
+        stride_handles[i] = 1;
+      } else {
+        stride_handles[i] = stride_handles[i + 1] * ExprHandle(dim(i + 1));
+      }
+    }
+    strides_ = ExprHandleVectorToExprVector(stride_handles);
+  }
+  Buffer(
+      const std::string& name,
+      const Dtype& dtype,
+      const std::vector<ExprHandle>& dims)
+      : Buffer(BufHandle(name, dims, dtype)) {}
+
+  const Buf* data() const {
+    return data_;
+  }
+  Dtype dtype() const {
+    return data_->dtype();
+  }
+  int ndim() const {
+    return data_->ndim();
+  }
+  const Expr* dim(int index) const {
+    return data_->dim(index);
+  }
+  std::vector<const Expr*> dims() const {
+    return data_->dims();
+  }
+
+  // TODO: consider defer the storage flatten to a later stage.
+  template <typename... Args>
+  ExprHandle operator()(Args... args) const {
+    return LoadValue(std::forward<Args>(args)...);
+  }
+
+  ExprHandle LoadValue(
+      const ExprHandle& x,
+      const ExprHandle& y,
+      const ExprHandle& z) const {
+    return Load::make(*this, {x, y, z}, ExprHandle(1));
+  }
+  ExprHandle LoadValue(const ExprHandle& x, const ExprHandle& y) const {
+    return Load::make(*this, {x, y}, ExprHandle(1));
+  }
+  ExprHandle LoadValue(const ExprHandle& x) const {
+    return Load::make(*this, {x}, ExprHandle(1));
+  }
+
+  template <typename T>
+  ExprHandle call(const std::vector<T>& args) const {
+    std::vector<ExprHandle> params(args.begin(), args.end());
+    return LoadValue(params);
+  }
+
+ private:
+  ExprHandle LoadValue(const std::vector<ExprHandle>& indices) const;
+
+  const Buf* data_;
+  std::vector<const Expr*> strides_;
+};
+
+inline ExprHandle Buffer::LoadValue(
+    const std::vector<ExprHandle>& indices) const {
+  return Load::make(*this, indices, ExprHandle(1));
+}
 
 TORCH_API Tensor* Compute(
     const std::string& func_name,
